+++ domain/src/main/java/net/boreeas/riotapi/com/riotgames/platform/game/GameMode.java
@@ -39,6 +39,6 @@ public enum GameMode {
             }
         }
 
-        return null;
+        throw new IllegalArgumentException("Unknown game mode: " + name);
     }
 }