+++ domain/src/main/java/net/boreeas/riotapi/com/riotgames/platform/summoner/MasteryGroup.java
@@ -39,4 +39,5 @@ public class MasteryGroup {
     private String name;
     @SerializedName(name = "tltGroupId")
     private int masteryGroupId;
+    private int version;
 }